requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.1

# LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# File parsers
pdfplumber>=0.10.3
python-docx>=1.1.0

# Reranking
sentence-transformers>=3.0.0
